+++ plugins/resources/nomiclabs.ui.js
@@ -1,7 +1,7 @@
 const UI = require('./../../lib/ui').UI;
 
 /**
- * Buidler Plugin logging
+ * Nomiclabs Plugin logging
  */
 class PluginUI extends UI {
   constructor(log){
@@ -36,18 +36,33 @@ class PluginUI extends UI {
       'instr-skip':  `\n${c.bold('Coverage skipped for:')}` +
                      `\n${c.bold('=====================')}\n`,
 
+      'compilation':  `\n${c.bold('Compilation:')}` +
+                      `\n${c.bold('============')}\n`,
+
       'instr-skipped': `${ds} ${c.grey(args[0])}`,
 
       'versions':  `${ct} ${c.bold('ganache-core')}:      ${args[0]}\n` +
                    `${ct} ${c.bold('solidity-coverage')}: v${args[1]}`,
 
-      'network': `\n${c.bold('Network Info')}` +
-                 `\n${c.bold('============')}\n` +
-                 `${ct} ${c.bold('port')}:    ${args[1]}\n` +
-                 `${ct} ${c.bold('network')}: ${args[0]}\n`,
+      'hardhat-versions': `\n${c.bold('Version')}` +
+                          `\n${c.bold('=======')}\n` +
+                          `${ct} ${c.bold('solidity-coverage')}: v${args[0]}`,
 
-      'port-clash': `${w}  ${c.red("The 'port' values in your Buidler url ")}` +
-                          `${c.red("and .solcover.js are different. Using Buidler's: ")} ${c.bold(args[0])}.\n`,
+      'hardhat-network': `\n${c.bold('Network Info')}` +
+                         `\n${c.bold('============')}\n` +
+                         `${ct} ${c.bold('HardhatEVM')}: v${args[0]}\n` +
+                         `${ct} ${c.bold('network')}:    ${args[1]}\n`,
+
+      'ganache-network': `\n${c.bold('Network Info')}` +
+                         `\n${c.bold('============')}\n` +
+                         `${ct} ${c.bold('port')}:         ${args[1]}\n` +
+                         `${ct} ${c.bold('network')}:      ${args[0]}\n`,
+
+      'port-clash': `${w}  ${c.red("The 'port' values in your config's network url ")}` +
+                          `${c.red("and .solcover.js are different. Using network's: ")} ${c.bold(args[0])}.\n`,
+
+      'port-clash-hardhat': `${w}  ${c.red("The 'port' values in your Hardhat network's url ")}` +
+                            `${c.red("and .solcover.js are different. Using Hardhat's: ")} ${c.bold(args[0])}.\n`,
 
     }
 
@@ -65,6 +80,8 @@ class PluginUI extends UI {
     const x = ":x:";
 
     const kinds = {
+      'network-fail': `${c.red('--network argument: ')}${args[0]}` +
+                      `${c.red(' is not a defined network in hardhat.config.js.')}`,
 
       'sources-fail': `${c.red('Cannot locate expected contract sources folder: ')} ${args[0]}`,